myenv/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
dnspython==2.5.0
psycopg2==2.9.9
pymongo==4.6.1
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
